Geological Model Age from Lead Isotope Analysis

Compute geological model age (T) and U/Pb (mu) and Th/U (kappa) ratios from lead isotopic measurements.

Usage

pb_age(x, y, z, ...)

# S4 method for numeric,numeric,numeric
pb_age(
  x,
  y,
  z,
  t0 = 3.8,
  x_star = 18.75,
  y_star = 15.63,
  z_star = 38.86,
  mu = 9.66,
  kappa = 3.9,
  th232 = 0.049475,
  u238 = 0.155125,
  u235 = 0.98485,
  u238_235 = 137.79,
  tolerance = sqrt(.Machine$double.eps)
)

# S4 method for list,missing,missing
pb_age(
  x,
  t0 = 3.8,
  x_star = 18.75,
  y_star = 15.63,
  z_star = 38.86,
  mu = 9.66,
  kappa = 3.9,
  th232 = 0.049475,
  u238 = 0.155125,
  u235 = 0.98485,
  u238_235 = 137.79,
  tolerance = sqrt(.Machine$double.eps)
)

Arguments

x

A numeric vector of 206Pb/204Pb ratios. If y and z are missing, must be a list (or a data.frame) with numeric components (columns) x, y and z.

y

A numeric vector of 207Pb/204Pb ratios. If missing, an attempt is made to interpret x in a suitable way.

z

A numeric vector of 208Pb/204Pb ratios. If missing, an attempt is made to interpret x in a suitable way.

...

Currently not used.

t0

A numeric value giving the time of the second stage of the reference model.

x_star

A numeric value giving the 206Pb/204Pb ratio at \(t = 0\).

y_star

A numeric value giving the 207Pb/204Pb ratio at \(t = 0\).

z_star

A numeric value giving the 208Pb/204Pb ratio at \(t = 0\).

mu

A numeric value giving the 238U/204Pb ratio of the reference model.

kappa

A numeric value giving the 232Th/238U ratio of the reference model.

th232

A numeric value giving the decay constants of 232Th.

u238

A numeric value giving the decay constants of 238U.

u235

A numeric value giving the decay constants of 235U.

u238_235

A numeric value giving the 238U/235U ratio.

tolerance

A numeric value specifying the tolerance (stopping criteria for the Newton–Raphson method).

Value

A four columns data.frame:

age

Geological model age (in Ma).

mu

238U/204Pb ratio.

kappa

232Th/238U ratio.

f

Note

Reference values from Albarede & Juteau (1984).

References

Albarède, F., Desaulty, A.-M. & Blichert-Toft, J. (2012). A Geological Perspective on the Use of Pb Isotopes in Archaeometry. Archaeometry, 54: 853-867. doi:10.1111/j.1475-4754.2011.00653.x .

Albarède, F. & Juteau, M. (1984). Unscrambling the Lead Model Ages. Geochimica et Cosmochimica Acta, 48(1): 207-12. doi:10.1016/0016-7037(84)90364-8 .

Author

N. Frerebeau, F. Albarede (original Matlab code)

Examples

Pb <- data.frame(
  x = c(18.23247, 18.22936, 18.23102), # Pb206/Pb204
  y = c(15.65199, 15.65216, 15.65097), # Pb207/Pb204
  z = c(38.5167, 38.51516, 38.51601)   # Pb208/Pb204
)

pb_age(Pb)
#>        age       mu    kappa            f
#> 1 428.3959 9.859142 4.074891 1.004859e-14
#> 2 430.9796 9.860659 4.076292 1.004859e-14
#> 3 427.6003 9.855397 4.075178 1.004859e-14