Plot Layer-Counted Proxy Records Uncertainties

Plot Layer-Counted Proxy Records Uncertainties

Usage

# S4 method for ProxyRecord,missing
plot(
  x,
  raw = FALSE,
  xlim = NULL,
  ylim = NULL,
  main = NULL,
  sub = NULL,
  ann = graphics::par("ann"),
  axes = TRUE,
  frame.plot = FALSE,
  panel.first = NULL,
  panel.last = NULL,
  ...
)

Arguments

x

A ProxyRecord object.

raw

A logical scalar: should the raw proxy data be displayed instead of the mean estimate?

xlim

A length-two numeric vector giving the x limits of the plot.

ylim

A length-two numeric vector giving the y limits of the plot.

main

A character string giving a main title for the plot.

sub

A character string giving a subtitle for the plot.

ann

A logical scalar: should the default annotation (title and x, y and z axis labels) appear on the plot?

axes

A logical scalar: should axes be drawn on the plot?

frame.plot

A logical scalar: should a box be drawn around the plot?

panel.first

An an expression to be evaluated after the plot axes are set up but before any plotting takes place. This can be useful for drawing background grids.

panel.last

An expression to be evaluated after plotting has taken place but before the axes, title and box are added.

...

Other graphical parameters may also be passed as arguments to this function.

Value

plot() is called it for its side-effects: it results in a graphic being displayed. Invisibly returns x.

See also

Author

N. Frerebeau

Examples

if (FALSE) {
## Get NGRIP records
data("ngrip2010", package = "folio")
ngrip2010 <- subset(ngrip2010, !is.na(MCE))

## Replicate fig. 3d from Boers et al. (2017)
ngrip_error <- proxy(
  depth = ngrip2010$depth,
  proxy = ngrip2010$delta,
  proxy_error = 0.01,
  time = ngrip2010$age,
  time_error = ngrip2010$MCE,
  grid = 0.01,
  resolution = 20,
  calendar = "b2k"
)

plot(ngrip_error, raw = FALSE, ylim = c(-46, -34))
}